# lib/model, lib/scanner, lib/weakhash: Single-pass weakhash.Finder