… go.sum
… lib/weakhash/benchmark_test.go
… lib/weakhash/weakhash.go
… lib/weakhash/weakhash_test.go