Successor to #6692.

This version uses a new [rolling hash library](/greatroar/rolling) that includes the file scanning logic and allows access to the scanner's buffer, so the weakhash.Finder is completely gone. I've moved its test to lib/model.

The scanner uses the fast non-rolling Adler-32 in this new library, which gives a small throughput boost:

```
name        old time/op    new time/op    delta
Validate-8    67.7ms ± 0%    65.8ms ± 0%  -2.81%  (p=0.000 n=9+10)
HashFile-8    87.7ms ± 0%    86.4ms ± 0%  -1.48%  (p=0.000 n=10+8)

name        old alloc/op   new alloc/op   delta
Validate-8     0.00B          0.00B         ~     (all equal)
HashFile-8    45.4kB ± 0%    45.4kB ± 0%    ~     (p=0.904 n=10+10)

name        old allocs/op  new allocs/op  delta
Validate-8      0.00           0.00         ~     (all equal)
HashFile-8      28.0 ± 0%      28.0 ± 0%    ~     (all equal)

name        old speed      new speed      delta
HashFile-8   203MB/s ± 0%   206MB/s ± 0%  +1.51%  (p=0.000 n=10+8)
```

The weakhash benchmarks are gone. The new library has its own benchmarks, which show that it can scan almost 60MB/s on my machine (more on a newer one, I suppose).