@imsodin 2020-07-24
I like the API and doing everything in a single pass. What makes me a bit concerned is that we move from an old, established one-person-show to another, new one-person-show. And move the scanning logic there. Which makes sense content-wise, but its still "our" code in that we need to debug it if there come up problems and that's not shared with someone else (unless you have more plans for it?). Ideally I'd have hoped @chmduquesne to react to your PR and get the changes in there. That ship has probably sailed though. One high-level question thus: Do you have any plans/expectations for that module outside of Syncthing and/or what's your motivation for creating a separate module?