@chmduquesne 2020-07-24
I can have a look later if you want. One thing, though: if you had complaints about my lib, why not file a ticket instead of creating an entirely new lib? I would favor joining forces, it makes the project stronger... The more maintainers the merrier! Happy to discuss.