@AudriusButkevicius 2020-07-24
Just an FYI, I am on holiday atm, but I plan to look at this in depth end of next week, given existing code I think was mostly my nonsense.

One comment, which perhaps other maintainers disagree with, is that we're adding quite a lot of dependencies on @greatroar repos in fairly sensitive areas, so I'm starting to be afraid we might blindly merge some dependabot PR bumping the dependency breaking the world for us, as the code is those projects is not reviewed by us