@imsodin 2020-07-24
@chmduquesne As far as I understand the main requirement for the previous iteration of this change (without a new lib) was https://github.com/chmduquesne/rollinghash/pull/10