@chmduquesne 2020-07-24
It seems like the API is quite different from my library, so I am not sure if I can really support you.

In the PR that I received, it looks like you are trying to access the content of the rolling window. It is true that there is an internal copy of this window kept in the rollinghash. This copy is continuously updated with each call to Roll(), in a circular buffer. Its only purpose is to know the value of the byte exiting the window: the idea of rollinghash was to avoid maintaining this buffer externally, because I found it annoying having to write h.Roll(enter, exit) when at a constant memory cost I could just write h.Roll(enter).

If you absolutely insist, I can add a way to extract this window (and add the WriteTo method suggested in chmduquesne/rollinghash#10). However, I am not sure this is very useful in the current context, since I don't see WriteTo used anywhere.