@chmduquesne 2020-07-24
By the way, I know that adler32 is popular because it is one of the first rolling hashes ever designed and it got traction thanks to rsync, but I would recommend switching to buzhash64 instead. It has less collisions for the same performance. If you are interested in knowing why, there is a little undocumented tool called roll (https://github.com/chmduquesne/rollinghash/blob/master/roll/main.go) which can be invoked with the option -stats. I wrote it to know how often I could expect a checksum beginning with a given number of 0s when hashing random data. If your checksum is well balanced, you should expect a checksum starting with n 0 should be expected every 2^n bytes. Buzshash64 works very well for that. In fact, every checksum in my package behaves well, *except* adler32. If you dig in the maths, it is obvious why and this behavior is expected, but then you should know that adler32 is really far from being the best rolling checksum.