@calmh 2020-07-24
@greatroar echoing the others, and adding personally, it would bring greater peace of mind when adding a dependency on a repo of yours if you had a name, or an email address, or had _existed anywhere_ longer than a few months. I don't see any problems with your contributions, but I also see no track record on which to base an assumption that these packages will be maintained in any fashion in the future. We can discuss these things offline, this may not be the correct forum, but it's a shared concern that has been brought up in the thread.