@greatroar 2020-07-27
I'll remove my lib and send patches to @chmduquesne instead.